Fixed #8 & NullPointerException, sockets closed
Fixed #8 & NullPointerException, sockets closed

- Fixed Issue #8. Obsolete code in ServerRpgCharater's observer managed
world serving. It is now being implemented by ServerGame RemoteClient's
UserHandler.

- Client Communicator did not properly respond to entities shared when a
world was not yet served. The client communicator would check if the
world was ready by invoking isReady on the world - which would throw a
NullPointerException if the world was not yet served. The communicator
now treats an unserved world as a unready world.

- m_world was marked nullable to prevent future misunderstanding when
dealing with it.

- The RemoteSocketCommunicator was not disconnecting its socket when it
had been unbound. RemoteCommunicators must now implement the abstract
methods onUnbind and onBind to handle the event in which they are bound
or unbound from their respective communicators. When the
RemoteSocketCommunicator is unbound, its respective socket is now
closed.

- Fixed some argument naming errors caused when a regex expression was
used to remove the existence of Hungarian Notation. Some might still
remain.

Jeremy. Allen. Wildsmith.