Resolved #4 & #5. Code Refactor & Javadoc Removal
Resolved #4 & #5. Code Refactor & Javadoc Removal

- WindowManager is now a service. This allows windows to be managed
outside of the game class; it also removes their dependency on the game
class. Resolves issue #4. If the core is not provided a window manager
with which to initialize, it will provide the BasicWindowManager as a
WindowManager service.

- Removed Javadoc. Was usually inaccurate and was incredibly
distracting. A new javadoc is will be introduced in later revisions.

- Resolved issue #5, Actors now have their commands handled and provided
by entity script. This change is currently only reflected in the zombie
script.

Jeremy. Allen. Wildsmith