Resolved #14 & Script Simplification, Res Cleanup
Resolved #14 & Script Simplification, Res Cleanup

- Refactored out EntityScript. EntityScript was inheriting script but
acting as an observer invoking script routines. I.e, the relation was
has-a, not is-a.

- Removed isReady, in Script since it was redundant.

- Scripts can now 'import' scripts or evaluate external scripts. This is
provided by the new variable 'script' accessible in scope of all loaded
scripts. The script variable exposes a ScriptContext object from which
you can invoke evaluateScript([script path])

- Due to the amount of Heap space consumed by a ScriptEngine, a
ScriptEngine is not actually initialized until it is evaluated to a
particular script.

- Enemy NPCs now inherit a 'common' NPC script to reduce boilerplate
code.

- Removed PARPG Assets.

Jeremy. Allen. Wildsmith